% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{build_indoor_spaces}
\alias{build_indoor_spaces}
\title{Build a new model that}
\usage{
build_indoor_spaces(enter_house, exit_house, time_indoors, model)
}
\arguments{
\item{enter_house}{a vector describing house entering rates}

\item{exit_house}{a vector describing house leaving rates}

\item{time_indoors}{fraction of time at risk spent indoors}

\item{model}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Buil
}
